Rename added component and formatting
